+++ CHANGES.md
@@ -9,6 +9,7 @@ Improvements 🙌:
  - Give user the possibility to prevent accidental call (#1869)
  - Display device information (name, id and key) in Cryptography setting screen (#1784)
  - Ensure users do not accidentally ignore other users (#1890)
+ - Stop background sync when network connection is lost (#1920)
 
 Bugfix 🐛:
  - Fix invisible toolbar (Status.im theme) (#1746)